Controller\Error -> Controller\Exception
Also moved MissingControllerException to Router, as it has a better home there
