+++ src/Controller/Exception/MissingComponentException.php
@@ -11,27 +11,16 @@
  * @since         3.0.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Controller\Error;
+namespace Cake\Controller\Exception;
 
 use Cake\Core\Exception\Exception;
 
 /**
- * Missing Action exception - used when a controller action
- * cannot be found.
+ * Used when a component cannot be found.
  *
  */
-class MissingActionException extends Exception {
+class MissingComponentException extends Exception {
 
-/**
- * {@inheritDoc}
- */
-	protected $_messageTemplate = 'Action %s::%s() could not be found.';
-
-/**
- * {@inheritDoc}
- */
-	public function __construct($message, $code = 404) {
-		parent::__construct($message, $code);
-	}
+	protected $_messageTemplate = 'Component class %s could not be found.';
 
 }
