+++ src/Controller/Exception/PrivateActionException.php
@@ -13,7 +13,7 @@
  * @since         3.0.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Controller\Error;
+namespace Cake\Controller\Exception;
 
 use Cake\Core\Exception\Exception;
