+++ tests/TestCase/Console/ConsoleErrorHandlerTest.php
@@ -15,7 +15,7 @@
 namespace Cake\Test\TestCase\Console;
 
 use Cake\Console\ConsoleErrorHandler;
-use Cake\Controller\Error\MissingActionException;
+use Cake\Controller\Exception\MissingActionException;
 use Cake\Core\Exception\Exception;
 use Cake\Log\Log;
 use Cake\Network\Exception\InternalErrorException;
