+++ tests/TestCase/Error/ExceptionRendererTest.php
@@ -16,10 +16,9 @@ namespace Cake\Test\TestCase\Error;
 
 use Cake\Controller\Component;
 use Cake\Controller\Controller;
-use Cake\Controller\Error\MissingActionException;
-use Cake\Controller\Error\MissingComponentException;
-use Cake\Controller\Error\MissingControllerException;
-use Cake\Controller\Error\PrivateActionException;
+use Cake\Controller\Exception\MissingActionException;
+use Cake\Controller\Exception\MissingComponentException;
+use Cake\Controller\Exception\PrivateActionException;
 use Cake\Core\App;
 use Cake\Core\Configure;
 use Cake\Core\Exception\MissingPluginException;
@@ -35,6 +34,7 @@ use Cake\Network\Exception\NotFoundException;
 use Cake\Network\Exception\SocketException;
 use Cake\Network\Request;
 use Cake\ORM\Error\MissingBehaviorException;
+use Cake\Routing\Error\MissingControllerException;
 use Cake\Routing\Router;
 use Cake\TestSuite\TestCase;
 use Cake\View\Error\MissingHelperException;
