+++ tests/TestCase/Routing/DispatcherTest.php
@@ -14,8 +14,6 @@
 namespace Cake\Test\TestCase\Routing;
 
 use Cake\Controller\Controller;
-use Cake\Controller\Error\MissingActionException;
-use Cake\Controller\Error\MissingControllerException;
 use Cake\Core\App;
 use Cake\Core\Configure;
 use Cake\Core\Plugin;
@@ -234,7 +232,7 @@ class DispatcherTest extends TestCase {
 /**
  * testMissingController method
  *
- * @expectedException \Cake\Controller\Error\MissingControllerException
+ * @expectedException \Cake\Routing\Error\MissingControllerException
  * @expectedExceptionMessage Controller class SomeController could not be found.
  * @return void
  */
@@ -253,7 +251,7 @@ class DispatcherTest extends TestCase {
 /**
  * testMissingControllerInterface method
  *
- * @expectedException \Cake\Controller\Error\MissingControllerException
+ * @expectedException \Cake\Routing\Error\MissingControllerException
  * @expectedExceptionMessage Controller class DispatcherTestInterface could not be found.
  * @return void
  */
@@ -273,7 +271,7 @@ class DispatcherTest extends TestCase {
 /**
  * testMissingControllerInterface method
  *
- * @expectedException \Cake\Controller\Error\MissingControllerException
+ * @expectedException \Cake\Routing\Error\MissingControllerException
  * @expectedExceptionMessage Controller class Abstract could not be found.
  * @return void
  */
